requirements.txt
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cryptography==36.0.1
joblib==1.1.0
nltk==3.6.7
numpy==1.22.0
pdfminer.six==20211012
pycparser==2.21
regex==2021.11.10
tqdm==4.62.3
